buyer/seller info fixes and minor changes
basically, i changed some text to be more concise.
also there are links connected to the words "create an account" so that newcomers don't have to navigate the menus, even.
finally, I thought the first time the three steps are stated were kind of redundant, so I commented them out; but if you want to you can get it back